requirements.txt
pandas
numpy
matplotlib
cudf-cu12
